added refresh action in home
added individual refresh progress on server cards

+++ app/src/main/java/net/myacxy/agsm/fragments/ServerDetailsFragment.java
@@ -24,7 +24,7 @@ public class ServerDetailsFragment extends Fragment
     ServerDetailsParameterAdapter serverDetailsParameterAdapter;
 
     @FragmentArg
-    int gameServerId;
+    long gameServerId;
 
     @AfterViews
     void initialize()
