+++ app/src/main/java/net/myacxy/agsm/views/adapters/ServerOverviewPlayersAdapter.java
@@ -32,7 +32,7 @@ public class ServerOverviewPlayersAdapter extends BaseAdapter
     @RootContext
     Context context;
 
-    private int gameServerId;
+    private long gameServerId;
 
     @Override
     public int getCount() {
@@ -90,7 +90,7 @@ public class ServerOverviewPlayersAdapter extends BaseAdapter
         }
     } // initAdapter
 
-    public void setGameServerId(int gameServerId)
+    public void setGameServerId(long gameServerId)
     {
         this.gameServerId = gameServerId;
         initAdapter();
